refactor(jobs): extract parseTransactionMetadata as a pure function
Pull the XDR-parsing logic out of TransactionStatusCheckerProcessor
into a standalone, throwing function so it can be unit tested in
isolation without constructing the full processor. The processor now
catches and logs the error itself, preserving the original graceful
fallback behavior. No public API change.
